+++ lib/main/ruby/SeleniumRuby.rb
@@ -1,3 +1,13 @@
+require "actions/Action"
+
+require "driver/Driver"
+
+require "test-cases/TestCase"
+
+require "test-suites/TestSuite"
+
+require "enums/Field"
+require "enums/Identifier"
+
 module SeleniumRuby
-    
 end
