+++ lib/main/ruby/test-cases/TestCase.rb
@@ -1,4 +1,4 @@
-class TestCases
+class TestCase
   def initialize(name)
     if name.nil? or name.empty?
       raise "All TestCasses must have a name associated to them"
